+++ test/string.js
@@ -198,11 +198,20 @@ describe('String', function() {
 	});
 
 	describe('#dissect(openTag, closeTag)', function() {
-		it('dissect a string into parts in- and outside of the given delimiters', function() {
+		it('should dissect a string into parts in- and outside of the given delimiters', function() {
 
 			var original = 'This <% is a %> test',
 			    arr      = original.dissect('<%', '%>'),
 			    expected = '[{"type":"normal","lineStart":0,"lineEnd":0,"content":"This "},{"type":"inside","lineStart":0,"lineEnd":0,"content":" is a "},{"type":"normal","lineStart":0,"lineEnd":0,"content":" test"}]';
+
+			assert.strictEqual(expected, JSON.stringify(arr));
+		});
+
+		it('should return an empty array for an empty string', function() {
+
+			var original = '',
+			    arr      = original.dissect('<%', '%>'),
+			    expected = '[]';
 
 			assert.strictEqual(expected, JSON.stringify(arr));
 		});