+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Turned URL#toString into a shim
 * Added Array#employ, for when Array#map is too verbose
 * Added Function.until, .doUntil and .while (sharing code with .doWhile)
+* BUGFIX: String#dissect will no longer throw an error on empty strings
 
 ## 0.1.3 (2014-09-15)